pyproject.toml
[tool.poetry]
name = "concertticket"
version = "0.1.0"
description = ""
authors = ["sayfullayevabdulaziz <abdulaziz13122002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
django-import-export = "^3.2.0"
django-jazzmin = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
